Extract DataGridViewTextBoxEditingControlAccessibleObject
