+++ src/System.Windows.Forms/src/System/Windows/Forms/DataGridViewTextBoxEditingControl.cs
@@ -9,7 +9,7 @@ using static Interop;
 
 namespace System.Windows.Forms
 {
-    public class DataGridViewTextBoxEditingControl : TextBox, IDataGridViewEditingControl
+    public partial class DataGridViewTextBoxEditingControl : TextBox, IDataGridViewEditingControl
     {
         private const DataGridViewContentAlignment AnyTop = DataGridViewContentAlignment.TopLeft | DataGridViewContentAlignment.TopCenter | DataGridViewContentAlignment.TopRight;
         private const DataGridViewContentAlignment AnyRight = DataGridViewContentAlignment.TopRight | DataGridViewContentAlignment.MiddleRight | DataGridViewContentAlignment.BottomRight;
@@ -308,107 +308,6 @@ namespace System.Windows.Forms
             {
                 _dataGridView?.SetAccessibleObjectParent(this.AccessibilityObject);
             }
-        }
-    }
-
-    /// <summary>
-    ///  Defines the DataGridView TextBox EditingControl accessible object.
-    /// </summary>
-    internal class DataGridViewTextBoxEditingControlAccessibleObject : Control.ControlAccessibleObject
-    {
-        private readonly DataGridViewTextBoxEditingControl ownerControl;
-
-        /// <summary>
-        ///  The parent is changed when the editing control is attached to another editing cell.
-        /// </summary>
-        private AccessibleObject _parentAccessibleObject;
-
-        public DataGridViewTextBoxEditingControlAccessibleObject(DataGridViewTextBoxEditingControl ownerControl) : base(ownerControl)
-        {
-            this.ownerControl = ownerControl;
-        }
-
-        public override AccessibleObject Parent
-        {
-            get
-            {
-                return _parentAccessibleObject;
-            }
-        }
-
-        public override string Name
-        {
-            get
-            {
-                string name = Owner.AccessibleName;
-                if (name != null)
-                {
-                    return name;
-                }
-                else
-                {
-                    return SR.DataGridView_AccEditingControlAccName;
-                }
-            }
-            set => base.Name = value;
-        }
-
-        internal override UiaCore.IRawElementProviderFragment FragmentNavigate(UiaCore.NavigateDirection direction)
-        {
-            switch (direction)
-            {
-                case UiaCore.NavigateDirection.Parent:
-                    if (Owner is IDataGridViewEditingControl owner && owner.EditingControlDataGridView.EditingControl == owner)
-                    {
-                        return _parentAccessibleObject;
-                    }
-
-                    return null;
-            }
-
-            return base.FragmentNavigate(direction);
-        }
-
-        internal override UiaCore.IRawElementProviderFragmentRoot FragmentRoot
-        {
-            get
-            {
-                return (Owner as IDataGridViewEditingControl)?.EditingControlDataGridView?.AccessibilityObject;
-            }
-        }
-
-        internal override object GetPropertyValue(UiaCore.UIA propertyID)
-        {
-            switch (propertyID)
-            {
-                case UiaCore.UIA.ControlTypePropertyId:
-                    return UiaCore.UIA.EditControlTypeId;
-                case UiaCore.UIA.NamePropertyId:
-                    return Name;
-                case UiaCore.UIA.IsValuePatternAvailablePropertyId:
-                    return true;
-            }
-
-            return base.GetPropertyValue(propertyID);
-        }
-
-        internal override bool IsPatternSupported(UiaCore.UIA patternId)
-        {
-            if (patternId == UiaCore.UIA.ValuePatternId)
-            {
-                return true;
-            }
-
-            return base.IsPatternSupported(patternId);
-        }
-
-        /// <summary>
-        ///  Sets the parent accessible object for the node which can be added or removed to/from hierachy nodes.
-        /// </summary>
-        /// <param name="parent">The parent accessible object.</param>
-        internal override void SetParent(AccessibleObject parent)
-        {
-            _parentAccessibleObject = parent;
         }
     }
 }
